feat(editor): refactor config browser and add migration preview test(editor): add unit tests for config browser and editor utilities
