+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2026-02-06
+
+**New**:
+- Added new unit tests for `ConfigTreeBuilder`, `ConfigValidationService`, and `ConfigsEditorUtil`
+
+**Changes**:
+- Decomposed monolithic `ConfigBrowserWindow` and `MigrationPanelElement` into MVC architecture
+- Simplified `README.md` by removing the inline package structure and contents
+
+---
+
 ## [1.0.0] - 2026-01-21
 
 **New**:
@@ -44,53 +55,53 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.7.0] - 2025-11-03
 
 **New**:
-- Added *Rebind* functionality to all Observable classes (*ObservableField*, *ObservableList*, *ObservableDictionary*) allowing rebinding to new data sources without losing existing observers
-- Added *Rebind* methods to all Observable Resolver classes (*ObservableResolverField*, *ObservableResolverList*, *ObservableResolverDictionary*) to rebind to new origin collections and resolver functions
-- Added new *IObservableResolverField* interface with *Rebind* method for resolver field implementations
+- Added `Rebind` functionality to all Observable classes (`ObservableField`, `ObservableList`, `ObservableDictionary`) allowing rebinding to new data sources without losing existing observers
+- Added `Rebind` methods to all Observable Resolver classes (`ObservableResolverField`, `ObservableResolverList`, `ObservableResolverDictionary`) to rebind to new origin collections and resolver functions
+- Added new `IObservableResolverField` interface with `Rebind` method for resolver field implementations
 
 ## [0.6.7] - 2025-04-07
 
 **New**:
-- Added the *UnityObjectExtensions* to help add extra logic to Unity's *GameObject* type objects
+- Added the `UnityObjectExtensions` to help add extra logic to Unity's `GameObject` type objects
 
 ## [0.6.6] - 2024-11-30
 
 **Fixes**:
-- *ObservableDictionary.Remove(T)* no longer sends an update if id doesn't find the element to remove it
+- `ObservableDictionary.Remove(T)` no longer sends an update if id doesn't find the element to remove it
 
 ## [0.6.5] - 2024-11-20
 
 **Fixes**:
-- Fixed the issues of *ObservableDictionary* when subscribing/unsubscribing to actions while removing/adding elements
-- Fixed the issues of *ObservableList* when subscribing/unsubscribing to actions while removing/adding elements
+- Fixed the issues of `ObservableDictionary` when subscribing/unsubscribing to actions while removing/adding elements
+- Fixed the issues of `ObservableList` when subscribing/unsubscribing to actions while removing/adding elements
 
 ## [0.6.4] - 2024-11-13
 
 **Fixes**:
-- Fixed the unit tests for *ObservableDictionary* that was preventing some builds to run
+- Fixed the unit tests for `ObservableDictionary` that was preventing some builds to run
 
 ## [0.6.3] - 2024-11-02
 
 **Fixes**:
-- Fixed the compilation issues of *ObservableDictionary*
+- Fixed the compilation issues of `ObservableDictionary`
 
 ## [0.6.2] - 2024-11-02
 
 **New**:
-- Added the *ObservableUpdateFlag* to help performance when updating subscribers to the *ObservableDictionary*. By default is set *ObservableUpdateFlag.KeyUpdateOnly*
+- Added the `ObservableUpdateFlag` to help performance when updating subscribers to the `ObservableDictionary`. By default is set `ObservableUpdateFlag.KeyUpdateOnly`
 
 **Fixes**:
-- Fixed an issue that would no setup Remove update action to Subscribers when calling *Clear* on the *ObservableDictionary*
+- Fixed an issue that would no setup Remove update action to Subscribers when calling `Clear` on the `ObservableDictionary`
 
 ## [0.6.1] - 2024-11-01
 
 **Fixes**:
-- Fixed an issue that would crash the execution when calling *Remove()* & *RemoveOrigin* from *ObservableResolverDictionary*
+- Fixed an issue that would crash the execution when calling `Remove()` & `RemoveOrigin` from `ObservableResolverDictionary`
 
 ## [0.6.0] - 2023-08-05
 
 **Changes**:
-- Improved the *ObservableResolverList* and *ObservableResolverDictionary* data types to properly resolve lists and dictionaries with different data types from the original collection.
+- Improved the `ObservableResolverList` and `ObservableResolverDictionary` data types to properly resolve lists and dictionaries with different data types from the original collection.
 
 ## [0.5.1] - 2023-09-04
 
@@ -120,13 +131,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.2.0] - 2020-09-28
 
 **New**:
-- Added new *ObservableResolverList*, *ObservableResolverDictionary* & *ObservableResolverField* to allow to create observable types without referencing the collection directly
+- Added new `ObservableResolverList`, `ObservableResolverDictionary` & `ObservableResolverField` to allow to create observable types without referencing the collection directly
 - Added Unit tests to all data types in the project
 
 **Changes**:
-- Removed *ObservableIdList* because it's behaviour was too confusing and the same result can be obtained with *ObservableList* or *ObservableDictionary*
-- Removed all Pair Data and moved them to new *Pair<Key,Value>* serialized type that can now be serializable on Unity 2020.1
-- Moved all Vector2, Vector3 & Vector4 extensions to the ValueData file
+- Removed `ObservableIdList` because it's behaviour was too confusing and the same result can be obtained with `ObservableList` or `ObservableDictionary`
+- Removed all Pair Data and moved them to new `Pair<Key,Value>` serialized type that can now be serializable on *Unity 2020.1*
+- Moved all `Vector2`, `Vector3` & `Vector4` extensions to the `ValueData` file
 
 ## [0.1.1] - 2020-08-31
