+++ Editor/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+// Copyright (c) GameLovers. All rights reserved.
+
+using System.Runtime.CompilerServices;
+
+/// <summary>
+/// Exposes internal editor types to the editor test assembly for unit testing.
+/// </summary>
+[assembly: InternalsVisibleTo("GameLovers.GameData.Editor.Tests")]
