+++ Editor/AssemblyInfo.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 153ac2b5b27004d94a039f8823fe1a94
