+++ Editor/Windows/ConfigBrowser.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f35d98a7c9a284a518ae8be84532a5ff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
