+++ Editor/Windows/ConfigBrowser/ConfigBrowserModels.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 05d72c9a7c079493383f75cbeeaaa2ca
