+++ Editor/Windows/ConfigBrowser/ConfigBrowserView.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 95fc0f1976ed740f19ca1dad1ed91372
