+++ Editor/Windows/ConfigBrowser/ConfigExportService.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d230fea222dc14fac818cb3f89fc7a5e
