+++ Editor/Windows/ConfigBrowser/ConfigTreeBuilder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2262662cc38584a07bf912d0671328be
