+++ Tests/Editor/Unit/ConfigBrowser.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8f1b97932cced4167bfd097133750cf2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
